+++ CHANGELOG.md
@@ -32,11 +32,12 @@ Attention: The newest changes should be on top -->
 
 ### Added
 
-- DOC: ASTRA Flight Example [#770](https://github.com/RocketPy-Team/RocketPy/pull/770))
+- DOC: ASTRA Flight Example [#770](https://github.com/RocketPy-Team/RocketPy/pull/770)
 
 ### Changed
 
 - MNT: add error handling for static margin calculation [#776](https://github.com/RocketPy-Team/RocketPy/pull/776)
+- MNT: EmptyMotor class inherits from Motor(ABC) [#779](https://github.com/RocketPy-Team/RocketPy/pull/779)
 
 ### Fixed